Fix client navigation not triggering content change
Previously, when somebody clicked a gatsby link on the page, the URL would change but the content would not. The `StoryblokEntry` component dictates which storyblok page to render. However, its state was never updated in response to a change in its props, which would occur when client side navigation changed the page to be rendered. As the state stayed the same, so would the page contents.